Shuffled the test to unit, since it's not actually hitting the service.
